Implement efficient 'caching' mechanism
Reading and writing file from/to disk is faster then
'sending/receiving' data. Therefore we simply create required images
in the /tmp directory. However, we only create, asynchronously, ath
the time they are required for viewing, instead of creating all images
fo the document at once (like doc-view). In this way we save memory as
typically we do not view all pages of a document in a single
'reading/emacs session'.